refactor: 패키징 수정 common service 분리
