+++ src/main/java/startwithco/startwithbackend/b2b/stat/domain/StatEntity.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import startwithco.startwithbackend.util.STAT_TYPE;
+import startwithco.startwithbackend.common.util.STAT_TYPE;
 import startwithco.startwithbackend.b2b.vendor.domain.VendorEntity;
 import startwithco.startwithbackend.base.BaseTimeEntity;
