+++ src/main/java/startwithco/startwithbackend/common/util/ENTRY_TYPE.java
@@ -1,4 +1,4 @@
-package startwithco.startwithbackend.util;
+package startwithco.startwithbackend.common.util;
 
 public enum ENTRY_TYPE {
     DEBIT,
