+++ src/main/java/startwithco/startwithbackend/common/util/SELL_TYPE.java
@@ -1,4 +1,4 @@
-package startwithco.startwithbackend.util;
+package startwithco.startwithbackend.common.util;
 
 public enum SELL_TYPE {
     SINGLE,
