+++ src/main/java/startwithco/startwithbackend/solution/erp/controller/request/ErpRequest.java
@@ -5,9 +5,9 @@ import org.springframework.util.CollectionUtils;
 import org.springframework.web.multipart.MultipartFile;
 import startwithco.startwithbackend.exception.badRequest.BadRequestErrorResult;
 import startwithco.startwithbackend.exception.badRequest.BadRequestException;
-import startwithco.startwithbackend.util.CATEGORY;
-import startwithco.startwithbackend.util.DIRECTION;
-import startwithco.startwithbackend.util.SELL_TYPE;
+import startwithco.startwithbackend.common.util.CATEGORY;
+import startwithco.startwithbackend.common.util.DIRECTION;
+import startwithco.startwithbackend.common.util.SELL_TYPE;
 
 import java.util.List;
