+++ src/main/java/startwithco/startwithbackend/solution/erp/repository/ErpEntityJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import startwithco.startwithbackend.solution.erp.domain.ErpEntity;
-import startwithco.startwithbackend.util.CATEGORY;
+import startwithco.startwithbackend.common.util.CATEGORY;
 
 import java.util.Optional;
