+++ src/main/java/startwithco/startwithbackend/solution/erp/repository/ErpEntityRepository.java
@@ -1,7 +1,7 @@
 package startwithco.startwithbackend.solution.erp.repository;
 
 import startwithco.startwithbackend.solution.erp.domain.ErpEntity;
-import startwithco.startwithbackend.util.CATEGORY;
+import startwithco.startwithbackend.common.util.CATEGORY;
 
 import java.util.Optional;
